mainSite/requirements.txt
asgiref==3.2.3
astroid==2.2.5
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
autopep8==1.4.4
certifi==2019.6.16
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-utils==0.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
joblib==0.14.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.18.2
oauthlib==3.0.1
pandas==1.0.3
pep8==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.3.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.1
requests==2.23.0
requests-oauthlib==1.2.0
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
zope.interface==4.7.2
